requirements.txt
WTForms~=2.3.3
PyYAML~=5.4.1
Flask~=2.0.1